clean up prerequests logic
